minor code, css clean up and included materialize

+++ client/app/account/conference/conference.controller.js
@@ -22,7 +22,7 @@ angular.module('sugarlandDoctorsApp')
 	    };
 
 	    var removeConference = function(conference) {
-			if(!$scope.isEndUser){
+			if(!$scope.isEndUser && conference){
 				$http.delete('/api/conferences/' + conference._id);
 			}
 	    };
